backend/requirements.txt
# FastAPI and ASGI server
fastapi
uvicorn[standard]

# Environment and API utilities
python-dotenv
requests
beautifulsoup4
alpha_vantage
newsapi-python

# LangChain and Vector DB
langchain
elasticsearch
langchain-elasticsearch
sentence-transformers
langchain-openai

# LangGraph and Multi-Agent System
langgraph
tavily-python 

# PyTorch and dependencies for NLP models
torch>=2.1.0
torchvision
torchaudio

# Document processing
PyPDF2
python-docx
python-multipart

langchain-google-genai
langchain_community